requirements.txt
fastapi
uvicorn
python-multipart
openai
tqdm
pypdf2
sentence-transformers
pinecone
python-dotenv
langchain-core
langchain-google-genai